requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
sentence-transformers
PyMuPDF
python-dotenv
groq
